Implements a centralized encoder register in the lib package. This allows all tools importing this library, to access encoding and decoding functionality for registered image type.
Sub-packages implement encoders for all supported
types. For some types, this is just a wrapper around
the encoders in the Go standard library, with the
inclusion of a generic encoder Option mechanism.

These options can be parsed from a string and can be
passed in through a command line parameter. They are
key/value pairs specifying values, specific to a given
image format's encoder.